fix: address CodeRabbit AI review feedback
